re-organised code, so that circuitpython documentation can automatically be generated
