+++ code/compare/compare.c
