--- code/extras.h
@@ -1,24 +1,0 @@
-
-/*
- * This file is part of the micropython-ulab project,
- *
- * https://github.com/v923z/micropython-ulab
- *
- * The MIT License (MIT)
- *
- * Copyright (c) 2020 Zoltán Vörös
-*/
-
-#ifndef _EXTRA_
-#define _EXTRA_
-
-#include "ulab.h"
-#include "ndarray.h"
-#include "fft.h"
-
-#if ULAB_EXTRAS_MODULE
-
-extern mp_obj_module_t ulab_extras_module;
-
-#endif
-#endif
