+++ code/linalg/linalg.c
@@ -161,9 +161,9 @@ MP_DEFINE_CONST_FUN_OBJ_1(linalg_inv_obj, linalg_inv);
 
 //| def dot(m1, m2):
 //|    """
-//|    :param ~ulab.array m1: a matrix
-//|    :param ~ulab.array m2: a matrix
-//|    Computes the matrix product of two matrices"""
+//|    :param ~ulab.array m1: a matrix, or a vector
+//|    :param ~ulab.array m2: a matrix, or a vector
+//|    Computes the product of two matrices, or two vectors. In the letter case, the inner product is returned."""
 //|
 //|    ...
 //|
