+++ code/micropython.mk
@@ -4,15 +4,15 @@ USERMODULES_DIR := $(USERMOD_DIR)
 # Add all C files to SRC_USERMOD.
 SRC_USERMOD += $(USERMODULES_DIR)/ndarray.c
 SRC_USERMOD += $(USERMODULES_DIR)/create.c
-SRC_USERMOD += $(USERMODULES_DIR)/linalg.c
-SRC_USERMOD += $(USERMODULES_DIR)/vectorise.c
-SRC_USERMOD += $(USERMODULES_DIR)/poly.c
-SRC_USERMOD += $(USERMODULES_DIR)/fft.c
-SRC_USERMOD += $(USERMODULES_DIR)/numerical.c
-SRC_USERMOD += $(USERMODULES_DIR)/filter.c
-SRC_USERMOD += $(USERMODULES_DIR)/compare.c
-SRC_USERMOD += $(USERMODULES_DIR)/approx.c
-SRC_USERMOD += $(USERMODULES_DIR)/extras.c
+SRC_USERMOD += $(USERMODULES_DIR)/linalg/linalg.c
+SRC_USERMOD += $(USERMODULES_DIR)/vector/vectorise.c
+SRC_USERMOD += $(USERMODULES_DIR)/poly/poly.c
+SRC_USERMOD += $(USERMODULES_DIR)/fft/fft.c
+SRC_USERMOD += $(USERMODULES_DIR)/numerical/numerical.c
+SRC_USERMOD += $(USERMODULES_DIR)/filter/filter.c
+SRC_USERMOD += $(USERMODULES_DIR)/compare/compare.c
+SRC_USERMOD += $(USERMODULES_DIR)/approx/approx.c
+SRC_USERMOD += $(USERMODULES_DIR)/user/user.c
 SRC_USERMOD += $(USERMODULES_DIR)/ulab.c
 
 CFLAGS_USERMOD += -I$(USERMODULES_DIR)
