+++ code/poly/poly.c
@@ -14,8 +14,7 @@
 #include "py/obj.h"
 #include "py/runtime.h"
 #include "py/objarray.h"
-#include "ndarray.h"
-#include "linalg.h"
+#include "../linalg/linalg.h"
 #include "poly.h"
 
 #if ULAB_POLY_MODULE
