+++ code/vector/vectorise.c
